Day #3 completed
Begun to fill out atmospherics center equipment, added some side areas next to telecomms. I'm not super sold on the design of telecomms right now so it might change in the future.

Today I finished:
-Marine Control Tower
-Telecomms
-Begun atmospherics work
-Crane Redesign
-Start of Accommodations section of the Rig (Admin, Research, CDC).

I was thinking of putting LCZ in that area as well but I might not have enough room so it's possible I'll end up having to extend the map size from 150 to something larger. Lots to work on!

Things I plan to work on:
-HCZ
-LCZ
-Accommodations
-Perhaps extending the Drill Floor?
-More control systems for the Marine Control tower.