05.07.24 (0:04)
Added some new directional water moving tiles for the tension legs when I get ready to do those.

Speaking of tension legs, got rid of the four ugly pillars and replaced it with two legs for tension winches. These will serve as the basis for the self-destruct system and I've moved them much closer to the central rig area so they're more accessible in an emergency.

Cleaned up the tiles in Logi Hallway, added Cargo Trolley (Might need to make a dedicated room and clean it up, it's really ugly looking).

Started work on AIC Core which I'll probably put in the Main Rig to the right of Logi.

I forget the other stuff I did.