Updates folders to newer versions to be buildable.
We really REALLY need to stop putting items in Map Folders. To run this build I have to duplicate everything in Site53 over to Site104 because the code is only built to support one map at the moment, and I haven't had the chance to clean it up or create my own map switching system.

This means every time there's a major update to the Site53 Map folder and things that need references get added there I have to update My Folder, or it's not going to compile.