A bunch of goodies! (7.11.24) (22:21)
Oh boy, It's been a minute since I updated it. So I wanted to fix up the main deck and make it feel more like one BIG oil deck rather then having holes everywhere, and so far I'm happy that I did. It's gonna be kind of weird near the edges but that's for me to figure out... *later*.

-Added Metal Fence code and sprites ported from TG which ported them from VG.
-Added Cargo Containers which are ported from TG which ported them from TGCM.
-Updated Turret code to recognize the security credentials as an access define that we actually use instead of ACCESS_SECURITY which we don't use anymore. (I use turrets in this map)
-Added Platform stairs ported from Aurora Station, I changed the sprites to have more holes in them since I wanted to have that industrial feel and their sprites felt a bit too... blocky, still working on the rest of them, right now I only fixed up the full one.

-That's about it, I also started adding some lights to other areas.