8.31.24
Cleaned up a lot of the side areas of the open deck, preparing to add the external maintenance areas to the map, gutted research (and put it floating above the map at the bottom as a sample for later) to make room for the AIC core which I want to be located in the North Rig. I felt AIC made more sense in the same area with Logi/HCZ/Engi than with CDC/LCZ/Research/Admin/Med

Toyed around with elevators but I might just gut that because why the hell can we not make double sided elevators??? I'll probably figure something out. (Shuttle Hijinks?)

Adjusted text for map info to be stuck in one piece rather than a bunch of unevenly separated text, I'll have to change it to whatever server directors want later anyway, so I might as well leave it this way for now.

Made the path from the South Rig to North Rig across the open deck more obvious, hopefully this should curb any confusion and keep a clear path while also allowing people to run off and find alternative routes if they want to.

Changed some other stuff that I really can't remember. Very Minor.