9.8.24 (18:10)
Started work on the AIC Core, good progress. Need to do the Roof, finish some interior hallway decorations, and map the entire electrical substation. I also need to map in the external catwalk that's going to serve as the auxiliary entrance if you're trying to be sneaky instead of going in through the checkpoint. I realized today that the Roof will also serve as a Third entry point so I'm thinking about how to make that interesting.

Fixed some minor double placed catwalks, patched some holes in reality, the usual.