September 29th, 2024
Tightened up the right side of the open deck and adjusted the frame of maintenance in that area so I can start fleshing it out later on. I might need to delay working on maintenance stuff so I can actually get the Main South Rig structure started. I keep slacking that because I want to make the North Rig more detailed.

Added a roof to the AIC Housing Area, and a general roof access starting point, will be fleshed out later on. So there's technically three AIC access points now.