10.5.24
Guts a lot of the deck for a redesign, I'm sick of all that plating. I want it even TIGHTER. Implements some small and large sprites for decoration and fun.

Implements a chunk of the breaker box code as a subset of APC just with a lid in front of it (creative I know). CanUseTopic still needs testing since I had to move it and haven't had the chance.

Modularizes a lot of APC code while I'm at it so people doing the same thing as me in the future will not have to go through as much as I did.