11.11.24
Fixed some Icons not showing in mapper because I forgot to update their default states.

Fleshed out CDZ a little more, added a lounge and medical clinic as well as some light maintenance areas and Two classrooms.

I plan to start working on departures or whatever I put in between CDZ and Departures next. It needs to be on the lowest level if I'm gonna make it a boat.