3.15.25
Implements small weather deck equipment monitoring room above research storage.

 Minor Roofing Fixes in that area.

Implements Auxillary Relay Tower and bridge leading up to Relay Tower (Floor-1 and start of 2 so far). The idea is to have an alternative long-distance satellite relay (bluespace relay used for emergency command messaging) setup on the top but have all of its systems off by default and perhaps have it deconstructed. It will serve as a location for survivors in an emergency to go and activate if the primary relay is unavailable, I intend to make it a multistep process that involves:

A. Activating the towers power grid, by either fueling the provided generators or connecting it to the main grid) and switching on all of the connected breaker boxes.

B. Constructing the relay itself with spare components that you brought or salvaged around the tower/deck for.

C. Someone having to send the message itself (But this is already an integral part of Command Communications, someone always has to have access to send the message, whether it's command staff, someone with their ID, or the AIC.

Perhaps something else physically in the tower but I can't think of anything right now.